Add optional `handle_error` function to Toniq.Worker module (#1)
* Add overridable 'handle_error' function to Worker
